+++ src/hotkeys.hh
@@ -34,5 +34,6 @@
 
 bool handle_keyseq(const char* keyseq);
 bool handle_paging_key(notcurses* nc, const ncinput& ch, const char* keyseq);
+void handle_paste_content(notcurses* nc, const ncinput& ch);
 
 #endif  // LNAV_HOTKEYS_H